brain/requirements.txt
brian2
pyzmq
numpy